feat: refine invited + compliance check components
